requirements.txt
Django==2.0.3
django-bootstrap-themes==3.3.6
django-bootstrap3==9.1.0
django-bootstrap4==0.0.6
django-graphos==0.3.41
pytz==2018.3
